pyproject.toml
[tool.poetry]
name = "example-discord-bot"
version = "0.1.0"
description = ""
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.4.0"
schemdraw = {extras = ["matplotlib"], version = "^0.19"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
